Nova_Layout_1-1
Fixed LED footprint, small pad is GND, large pad is positive. This is opposite of the data sheet.

Adjusted position of some components for better fit (and fixed R48 and R49)

Switched the input pins for the transistors to fit a 3x1 set of pins with 2.54mm pitch.

Changed inner resistor and outer resistor footprints to match original outer footprints.

Checked all reference numbers to make sure they were the correct size and will be readable when soldering.